# PMP: Self_intersections_with_tree