## Summary of Changes

Use a custom traaits for aabb_tree to compute self_intersections.

Bench [here](https://gist.github.com/maxGimeno/fdea8214214b624a2fff4385d7388d81)
## Release Management

* Affected package(s):PMP
* Feature/Small Feature (if any):
* Link to compiled documentation (obligatory for small feature) [*wrong link name to be changed*](httpssss://wrong_URL_to_be_changed/Manual/Pkg)
* License and copyright ownership: