@sloriot 2019-11-07
The timing are not favorable to this implementation because `Box_intersection_d` test box vs box instead of face vs tree in the current implementation. I guess we have to use the tree/tree intersection.