@sloriot 2019-11-07
For archive my timing with 289ce74 

Input has 703512 faces.
Building the tree takes: 0.451887s.
New version in parallel takes: 15.6277s (0).
New version in sequential takes: 90.8133s (0).
Old version takes: 1.21053s (0).